SEQUOIADBMAINSTREAM-3946 修复om插件不支持无符号整形抛异常的问题
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37804 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ tools/om_plugins/sequoiasql/source/src/main/java/com/sequoiadb/om/plugin/dao/SequoiaSQLOperations.java
@@ -101,16 +101,16 @@ public abstract class SequoiaSQLOperations {
                     json.put(columnName, rs.getArray(i));
                     break;
                 case Types.BIGINT:
-                    json.put(columnName, rs.getLong(i));
+                    json.put(columnName, rs.getBigDecimal(i));
                     break;
                 case Types.INTEGER:
-                    json.put(columnName, rs.getInt(i));
+                    json.put(columnName, rs.getBigDecimal(i));
                     break;
                 case Types.TINYINT:
-                    json.put(columnName, rs.getByte(i));
+                    json.put(columnName, rs.getBigDecimal(i));
                     break;
                 case Types.SMALLINT:
-                    json.put(columnName, rs.getShort(i));
+                    json.put(columnName, rs.getBigDecimal(i));
                     break;
                 case Types.BIT:
                 case Types.BOOLEAN:
